move FullWidthErrorMsg to the components folder
