+++ src/pages/ResetPasswordStep2.js
@@ -6,7 +6,7 @@ import useFormField from "../hooks/useFormField";
 
 import Form from "./_pages_shared/Form";
 import FormSection from "./_pages_shared/FormSection";
-import FullWidthErrorMsg from "./_pages_shared/FullWidthErrorMsg";
+import FullWidthErrorMsg from "../components/FullWidthErrorMsg";
 import InputField from "../components/InputField";
 import ButtonContainer from "./_pages_shared/ButtonContainer";
 import Button from "./_pages_shared/Button";
