+++ src/pages/Settings/Classrooms.js
@@ -11,7 +11,7 @@ import Header from "../_pages_shared/Header";
 import Heading from "../_pages_shared/Heading";
 import Main from "../_pages_shared/Main";
 import { FullWidthListContainer } from "./settings_pages_shared/FullWidthListContainer.sc";
-import FullWidthErrorMsg from "../_pages_shared/FullWidthErrorMsg";
+import FullWidthErrorMsg from "../../components/FullWidthErrorMsg";
 import BackArrow from "./settings_pages_shared/BackArrow";
 import strings from "../../i18n/definitions";
 import LoadingAnimation from "../../components/LoadingAnimation";
