+++ src/pages/Settings/Languages.js
@@ -15,7 +15,7 @@ import PreferencesPage from "../_pages_shared/PreferencesPage";
 import Header from "../_pages_shared/Header";
 import Heading from "../_pages_shared/Heading";
 import Main from "../_pages_shared/Main";
-import FullWidthErrorMsg from "../_pages_shared/FullWidthErrorMsg";
+import FullWidthErrorMsg from "../../components/FullWidthErrorMsg";
 import BackArrow from "./settings_pages_shared/BackArrow";
 import Selector from "../../components/Selector";
