+++ src/pages/SignIn.js
@@ -8,7 +8,7 @@ import Heading from "./_pages_shared/Heading";
 import Main from "./_pages_shared/Main";
 import Form from "./_pages_shared/Form";
 import FormSection from "./_pages_shared/FormSection";
-import FullWidthErrorMsg from "./_pages_shared/FullWidthErrorMsg";
+import FullWidthErrorMsg from "../components/FullWidthErrorMsg";
 import InputField from "../components/InputField";
 import Footer from "./_pages_shared/Footer";
 import ButtonContainer from "./_pages_shared/ButtonContainer";
